--- helm/kanister-operator/crds/repositoryserver.yaml
@@ -1,1 +1,0 @@
-../../../pkg/customresource/repositoryserver.yaml
