+++ pkg/apis/cr/v1alpha1/doc.go
@@ -1,5 +1,5 @@
 // +k8s:deepcopy-gen=package,register
 
 // Package v1alpha1 is the v1alpha1 version of the API.
-// +groupName=cr.kanister.io
+// +groupName=cr
 package v1alpha1
