+++ pkg/apis/cr/v1alpha1/types.go
@@ -22,7 +22,6 @@ which also has the apache 2.0 license.
 package v1alpha1
 
 import (
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -375,111 +374,4 @@ type ProfileList struct {
 	metav1.ListMeta `json:"metadata"`
 	// Items represents a list of Profiles.
 	Items []*Profile `json:"items"`
-}
-
-// +genclient
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-// +kubebuilder:subresource:status
-
-// RepositoryServer manages the lifecycle of Kopia Repository Server within a Pod
-type RepositoryServer struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	Spec              RepositoryServerSpec   `json:"spec"`
-	Status            RepositoryServerStatus `json:"status"`
-}
-
-// RepositoryServerSpec is the specification for the RepositoryServer
-type RepositoryServerSpec struct {
-	Storage    Storage    `json:"storage"`
-	Repository Repository `json:"repository"`
-	Server     Server     `json:"server"`
-}
-
-// Storage references the backend store where a repository already exists
-// and the credential necessary to connect to the backend store
-type Storage struct {
-	SecretRef           corev1.SecretReference `json:"secretRef"`
-	CredentialSecretRef corev1.SecretReference `json:"credentialSecretRef"`
-}
-
-// Repository details for the purpose of establishing a connection
-type Repository struct {
-	RootPath          string                 `json:"rootPath"`
-	Username          string                 `json:"username"`
-	Hostname          string                 `json:"hostname"`
-	PasswordSecretRef corev1.SecretReference `json:"passwordSecretRef"`
-}
-
-// Server details required for starting the repository proxy server and initializing the repository client users
-type Server struct {
-	UserAccess     UserAccess             `json:"userAccess"`
-	AdminSecretRef corev1.SecretReference `json:"adminSecretRef"`
-	TLSSecretRef   corev1.SecretReference `json:"tlsSecretRef"`
-}
-
-type UserAccess struct {
-	UserAccessSecretRef corev1.SecretReference `json:"userAccessSecretRef"`
-	Username            string                 `json:"username"`
-}
-
-// RepositoryServerStatus is the status for the RepositoryServer. This should only be updated by the controller
-type RepositoryServerStatus struct {
-	Conditions []Condition              `json:"conditions,omitempty" patchStrategy:"merge" patchMergeKey:"type"`
-	ServerInfo ServerInfo               `json:"serverInfo,omitempty"`
-	Progress   RepositoryServerProgress `json:"progress"`
-}
-
-// Condition contains details of the current state of the RepositoryServer resource
-type Condition struct {
-	LastTransitionTime metav1.Time                   `json:"lastTransitionTime,omitempty"`
-	LastUpdateTime     metav1.Condition              `json:"lastUpdateTime,omitempty"`
-	Status             metav1.ConditionStatus        `json:"status"`
-	Type               RepositoryServerConditionType `json:"type"`
-}
-
-// RepositoryServerConditionType defines all the various condition types of the RepositoryServer resource
-type RepositoryServerConditionType string
-
-const (
-	// RepositoryReady indicates whether the existing repository is connected and ready to use
-	RepositoryReady RepositoryServerConditionType = "RepositoryReady"
-
-	// ServerInitialized means that the proxy server, that serves the repository, has been started
-	ServerInitialized RepositoryServerConditionType = "ServerInitialized"
-
-	// ClientsInitialized indicates that the client users have been added or updated to the repository server
-	ClientsInitialized RepositoryServerConditionType = "ClientsInitialized"
-
-	// ServerRefreshed denotes the refreshed condition of the repository server in order to register client users
-	ServerRefreshed RepositoryServerConditionType = "ServerRefreshed"
-)
-
-// RepositoryServerProgress is the field users would check to know the state of RepositoryServer
-type RepositoryServerProgress string
-
-const (
-	// ServerReady represents the ready state of the repository server and the pod which runs the proxy server
-	ServerReady RepositoryServerProgress = "ServerReady"
-
-	// ServerStopped represents the terminated state of the repository server pod due to any unforeseen errors
-	ServerStopped RepositoryServerProgress = "ServerStopped"
-
-	// ServerPending indicates the pending state of the RepositoryServer CR when Reconcile callback is in progress
-	ServerPending RepositoryServerProgress = "ServerPending"
-)
-
-// ServerInfo describes all the information required by the client users to connect to the repository server
-type ServerInfo struct {
-	PodName     string `json:"podName,omitempty"`
-	ServiceName string `json:"serviceName,omitempty"`
-}
-
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-// RepositoryServerList is the definition of a list of RepositoryServers
-type RepositoryServerList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []RepositoryServer `json:"items"`
 }
